+++ src/ces/export.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <iostream>
+#include "parser.h"
+
+typedef struct yy_buffer_state* YY_BUFFER_STATE;
+extern YY_BUFFER_STATE yy_scan_string(const char* str);
+extern void yy_delete_buffer(YY_BUFFER_STATE buffer);
+
+BaseNode* ces(std::string src);