+++ test/test.ts
@@ -1,4 +1,5 @@
 import { node_windows_x64 as nw } from "../";
 
-nw.getAddress(`1^2`);
+nw.registerSymbol("中文", 2);
 
+console.log( nw.getAddress("中文+2") );